split AccountAiTab in multiples files
