+++ web/src/ui/i18n/resources/de.tsx
@@ -110,18 +110,39 @@ export const translations: Translations<"de"> = {
         ),
         "api base url": "API-Basis-URL",
         token: "Token",
-        "model label": "Modell",
         "gateway error": "Das KI-Gateway konnte nicht initialisiert werden.",
         "default provider": "Standardanbieter",
         "set default provider": "Als Standard festlegen",
         "refresh credentials": "Anmeldedaten aktualisieren",
         "delete provider": "Löschen",
         "edit provider": "Bearbeiten",
-        copy: "Kopieren",
-        "not defined": "Nicht definiert",
         "custom providers section title": "Benutzerdefinierte KI-Anbieter",
         "custom providers section helper":
             "Fügen Sie Ihre eigenen OpenAI-kompatiblen KI-Anbieter hinzu. Die Anmeldedaten werden in Ihrem Browser gespeichert.",
+        "custom provider api base field": "API-Basis-URL",
+        "custom provider api key field": "API-Schlüssel",
+        "no account": ({ webUiUrl }) => (
+            <>
+                Sie haben noch kein Konto beim KI-Gateway. Bitte melden Sie sich zuerst an
+                bei{" "}
+                <MuiLink href={webUiUrl} target="_blank">
+                    {webUiUrl}
+                </MuiLink>{" "}
+                um Ihr Konto zu erstellen.
+            </>
+        )
+    },
+    ProviderValueField: {
+        copy: "Kopieren"
+    },
+    ModelsSection: {
+        "model label": "Modell",
+        "not defined": "Nicht definiert",
+        "models fetch error":
+            "Modelle konnten nicht abgerufen werden — überprüfen Sie URL und API-Schlüssel."
+    },
+    CustomProviderFormDialog: {
+        "add custom provider title": "Benutzerdefinierte KI-Anbieter",
         "edit custom provider title": "KI-Anbieter bearbeiten",
         "custom provider label field": "Name",
         "custom provider type field": "Provider-Typ",
@@ -133,19 +154,7 @@ export const translations: Translations<"de"> = {
             "Verbindung fehlgeschlagen — URL und API-Schlüssel prüfen.",
         "provider save": "Hinzufügen",
         "provider update": "Speichern",
-        "provider cancel": "Abbrechen",
-        "models fetch error":
-            "Modelle konnten nicht abgerufen werden — überprüfen Sie URL und API-Schlüssel.",
-        "no account": ({ webUiUrl }) => (
-            <>
-                Sie haben noch kein Konto beim KI-Gateway. Bitte melden Sie sich zuerst an
-                bei{" "}
-                <MuiLink href={webUiUrl} target="_blank">
-                    {webUiUrl}
-                </MuiLink>{" "}
-                um Ihr Konto zu erstellen.
-            </>
-        )
+        "provider cancel": "Abbrechen"
     },
     AccountVaultTab: {
         "credentials section title": "Vault-Anmeldeinformationen",
