+++ web/src/ui/i18n/types.ts
@@ -55,7 +55,10 @@ export type ComponentKey =
     | import("ui/pages/account/AccountKubernetesTab").I18n
     | import("ui/pages/account/AccountUserInterfaceTab").I18n
     | import("ui/pages/account/AccountVaultTab").I18n
-    | import("ui/pages/account/AccountAiTab").I18n
+    | import("ui/pages/account/AccountAiTab/AccountAiTab").I18n
+    | import("ui/pages/account/AccountAiTab/ProviderValueField").I18n
+    | import("ui/pages/account/AccountAiTab/ModelsSection").I18n
+    | import("ui/pages/account/AccountAiTab/CustomProviderFormDialog").I18n
     | import("ui/App/Footer").I18n
     | import("ui/pages/catalog/Page").I18n
     | import("ui/pages/catalog/CatalogChartCard").I18n
