+++ web/src/ui/pages/account/AccountAiTab/index.ts
@@ -1,0 +1,3 @@
+import { AccountAiTab } from "./AccountAiTab";
+
+export default AccountAiTab;
